Create pyspark.sql.connect.utils to keep common codes
